+++ RELEASES.md
@@ -1,5 +1,25 @@
 # Releases
 
+## 1.3.0 - 2025-10-25
+
+**5 commits** from **1 contributors**
+
+Changes since v0.2.3:
+
+**Highlights:**
+- 5 other changes
+
+**Key Changes:**
+- v1.3.0 renamed grpt
+- 📚 Updated tool documentation and added git information flag
+- 🚀 Added external config
+- 🔧 Fixed prefix 'v' issue in version manager and updated version to v0.2.3
+- 🔧 Update version to v0.2.2 across all projects, add known bugs section to README, extend versioning patterns for specific files
+
+**Contributors:** paulrm
+
+**Commit Hash:** `3d9b4f3`
+
 ## v0.1.1 - 2025-09-26
 
 **1 commits** from **1 contributors**